Fixed v2 tests
