+++ kpi/views/v1/hook_signal.py
@@ -32,4 +32,4 @@ class HookSignalViewSet(HookSignalViewSetV2):
 
     """
 
-    URL_NAMESPACE = None
+    pass
